restructured repo
renamed tictactoe.py to tictactoe_board.py
moved tictactoe_board.py to GUI module
refactored GameState to gamestate.py
moved minimax.py and gamestate.py into minimax module
